Create Readme
Differential expression analysis was performed using edgeR package in R on the data from BS patients, TS patients, and healthy controls. The screening threshold for differentially expressed mRNAs (DEmRs) and lncRNAs (DElncRs) is set at |log2 fold change| > 1 and P < 0.05.
Subsequently, we used the Weighted Gene Coexpression Network Analysis (WGCNA) method to construct a co-expression network. Firstly, we calculated the correlation between differentially expressed genes based on their expression levels in samples. Then, we selected the minimum β value to make the weighted network fit the scale-free network. We used the concept of 'module' to represent a group of highly co-expressed genes. Then, we conducted correlation analysis between the module eigengenes of each module and the clinical traits of samples to identify the modules that are significantly associated with clinical traits.
In the enrichment analysis, we used clusterProfiler package in R to perform enrichment analysis of Kyoto Encyclopedia of Genes and Genomes (KEGG) on DEmRs and module genes. P < 0.05 was considered as the threshold for significant enrichment.